git-svn-id: http://www.mz3.jp/svn/mz3/trunk@1616 875e8a4c-fb2d-0410-984e-92a9dd373959

+++ src/MZ3/BookmarkMng.cpp
@@ -159,7 +159,7 @@ BOOL Bookmark::Add( CMixiData* master, CMixiDataList& bookmark )
 
 	CMixiData data;
 	data.SetName(master->GetName());
-	data.SetID(master->GetID());
+	data.SetID((int)master->GetID());
 	data.SetAccessType(master->GetAccessType());
 	data.SetCommentCount(master->GetCommentCount());
 	data.SetURL(master->GetURL());
